back_course/requirements.txt
# requirements.txt
fastapi[standart]==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9